Pruning dependencies, cleaning code

+++ src/main/java/edu/virginia/lib/ole/akubra/Constants.java
@@ -6,6 +6,6 @@ package edu.virginia.lib.ole.akubra;
  */
 public final class Constants {
 	
-	public static Byte CHUNK = 2;
+	public static byte CHUNK = (byte)2;
 
 }
